JSON slightly better defined, but still pretty bad
JSON user data is now optional (cool stuff), might change name or id to be required.

function ProcessStandardUserData removed and slightly replaced by CheckAndDefineData.

CheckAndDefineData takes data, checks all possible variables in the data if they have actual values, returns true if any data was undefined and also changes the data so it is "defined"... will probably change this in the future so it won't. hmm ;/

MusicCommand, !play now does not automatically move the bot to the voicechannel. client.voiceConnections is also interesting me cause it suggests the client can be in several voice channels at the same time, therefore the code plays sound in all voice connections the client has. But so far it can only be in one, which seems to be the only choice...?